+++ src/rtp.h
@@ -33,8 +33,8 @@ enum encoding {
 };
 
 struct pt_table {
-  unsigned int samprate;
-  unsigned int channels;
+  int samprate;
+  int channels;
   enum encoding encoding;
 };
 
@@ -121,7 +121,7 @@ struct packet {
 void const *ntoh_rtp(struct rtp_header *,void const *);
 void *hton_rtp(void *, struct rtp_header const *);
 
-int add_pt(int type, unsigned int samprate, unsigned int channels, enum encoding encoding);
+int add_pt(int type, int samprate, int channels, enum encoding encoding);
 // Function to process incoming RTP packet headers
 // Returns number of samples dropped or skipped by silence suppression, if any
 int rtp_process(struct rtp_state *state,struct rtp_header const *rtp,size_t samples);
@@ -155,7 +155,7 @@ bool legal_opus_samprate(int n);
 int samprate_from_pt(int type);
 int channels_from_pt(int type);
 enum encoding encoding_from_pt(int type);
-uint8_t pt_from_info(unsigned int samprate,unsigned int channels,enum encoding);
+int pt_from_info(int samprate,int channels,enum encoding);
 char const *encoding_string(enum encoding);
 enum encoding parse_encoding(char const *str);
